cleanup expose-strategy tests
