+++ hack/run-expose-strategy-e2e-test-in-kind.sh
@@ -205,7 +205,6 @@ echodate "Kubermatic ingress domain patched."
 
 echodate "Running tests..."
 go test -tags "$KUBERMATIC_EDITION,e2e" -v ./pkg/test/e2e/expose-strategy \
-  -ginkgo.v \
   -kubeconfig "$HOME/.kube/config" \
   -kubernetes-version "$USER_CLUSTER_KUBERNETES_VERSION" \
   -datacenter byo-kubernetes \
