+++ pkg/test/e2e/expose-strategy/tunneling_test.go
@@ -20,73 +20,102 @@ package exposestrategy
 
 import (
 	"context"
-
-	"github.com/onsi/ginkgo"
-	"github.com/onsi/gomega"
+	"flag"
+	"testing"
 
 	"k8c.io/kubermatic/v2/pkg/log"
+	"k8c.io/kubermatic/v2/pkg/semver"
 	e2eutils "k8c.io/kubermatic/v2/pkg/test/e2e/utils"
 	"k8c.io/kubermatic/v2/pkg/version/kubermatic"
 
 	"k8s.io/apimachinery/pkg/util/rand"
 )
 
-var _ = ginkgo.Describe("The Tunneling strategy", func() {
+// Options holds the e2e test options.
+type testOptions struct {
+	skipCleanup       bool
+	logOptions        log.Options
+	datacenter        string
+	kubernetesVersion semver.Semver
+}
+
+var (
+	options = testOptions{
+		kubernetesVersion: semver.Semver(e2eutils.KubernetesVersion()),
+		logOptions:        e2eutils.DefaultLogOptions,
+	}
+)
+
+func init() {
+	flag.StringVar(&options.datacenter, "datacenter", "byo-kubernetes", "Name of the datacenter used by the user clusters created for the test.")
+	flag.Var(&options.kubernetesVersion, "kubernetes-version", "Kubernetes version for the user cluster")
+	flag.BoolVar(&options.skipCleanup, "skip-cleanup", false, "Skip clean-up of resources")
+
+	options.logOptions.AddFlags(flag.CommandLine)
+}
+
+func TestExposeKubernetesApiserver(t *testing.T) {
 	ctx := context.Background()
 	logger := log.NewFromOptions(options.logOptions).Sugar()
+	k8scli, restCli, restConf := e2eutils.GetClientsOrDie()
 
-	var (
-		clusterJig  *ClusterJig
-		agentConfig *AgentConfig
-		client      *clientJig
-	)
-	ginkgo.BeforeEach(func() {
-		k8scli, restCli, restConf := e2eutils.GetClientsOrDie()
-		clusterJig = &ClusterJig{
-			Log:            logger,
-			Client:         k8scli,
-			Name:           "c" + rand.String(9),
-			DatacenterName: options.datacenter,
-			Version:        options.kubernetesVersion,
-		}
-		gomega.Expect(clusterJig.SetUp(ctx)).NotTo(gomega.HaveOccurred(), "user cluster should deploy successfully")
-		agentConfig = &AgentConfig{
-			Log:       logger,
-			Client:    k8scli,
-			Namespace: clusterJig.Cluster.Status.NamespaceName,
-			Versions:  kubermatic.NewDefaultVersions(),
-		}
-		client = &clientJig{e2eutils.TestPodConfig{
-			Log:           logger,
-			Namespace:     clusterJig.Cluster.Status.NamespaceName,
-			Client:        k8scli,
-			PodRestClient: restCli,
-			Config:        restConf,
-			CreatePodFunc: newClientPod,
-		}}
-		gomega.Expect(agentConfig.DeployAgentPod(ctx)).NotTo(gomega.HaveOccurred(), "agent should deploy successfully")
-		gomega.Expect(client.DeployTestPod(ctx, logger)).NotTo(gomega.HaveOccurred(), "client pod should deploy successfully")
-	})
-	ginkgo.AfterEach(func() {
-		if !options.skipCleanup {
-			gomega.Expect(clusterJig.CleanUp(ctx)).NotTo(gomega.HaveOccurred())
-			if agentConfig != nil {
-				gomega.Expect(agentConfig.CleanUp(ctx)).NotTo(gomega.HaveOccurred())
+	clusterJig := &ClusterJig{
+		Log:            logger,
+		Client:         k8scli,
+		Name:           "c" + rand.String(9),
+		DatacenterName: options.datacenter,
+		Version:        options.kubernetesVersion,
+	}
+	if err := clusterJig.SetUp(ctx); err != nil {
+		t.Errorf("Failed to setup usercluster: %v", err)
+		return
+	}
+
+	agentConfig := &AgentConfig{
+		Log:       logger,
+		Client:    k8scli,
+		Namespace: clusterJig.Cluster.Status.NamespaceName,
+		Versions:  kubermatic.NewDefaultVersions(),
+	}
+	if err := agentConfig.DeployAgentPod(ctx); err != nil {
+		t.Errorf("Failed to deploy agent: %v", err)
+		return
+	}
+
+	client := &clientJig{e2eutils.TestPodConfig{
+		Log:           logger,
+		Namespace:     clusterJig.Cluster.Status.NamespaceName,
+		Client:        k8scli,
+		PodRestClient: restCli,
+		Config:        restConf,
+		CreatePodFunc: newClientPod,
+	}}
+	if err := client.DeployTestPod(ctx, logger); err != nil {
+		t.Errorf("Failed to deploy Pod: %v", err)
+		return
+	}
+
+	logger.Debug("Testing SNI when Kubeconfig is used e.g. Kubelet")
+	if result := client.QueryApiserverVersion("", false, options.kubernetesVersion, 5, 4); result != true {
+		t.Error("Apiserver should be reachable passing from the SNI entrypoint in nodeport proxy")
+	}
+
+	logger.Debug("Tunneling requests using HTTP/2 CONNECT when no SNI is present e.g. pods relying on kubernetes service in default namespace")
+	if result := client.QueryApiserverVersion(agentConfig.GetKASHostPort(), true, options.kubernetesVersion, 5, 4); result != true {
+		t.Error("Apiserver should be reachable passing from the SNI entrypoint in nodeport proxy")
+	}
+
+	if !options.skipCleanup {
+		defer func() {
+			if err := client.CleanUp(ctx); err != nil {
+				t.Errorf("Failed to cleanup: %v", err)
 			}
-			if client != nil {
-				gomega.Expect(client.CleanUp(ctx)).NotTo(gomega.HaveOccurred())
+			if err := agentConfig.CleanUp(ctx); err != nil {
+				t.Errorf("Failed to cleanup: %v", err)
 			}
-		}
-	})
-	ginkgo.Context("uses nodeport-proxy to expose control plane components", func() {
-		ginkgo.It("exposes the KAS", func() {
-			ginkgo.By("relying on SNI when Kubeconfig is used e.g. Kubelet")
-			gomega.Expect(client.QueryApiserverVersion("", false, options.kubernetesVersion, 5, 4)).To(gomega.BeTrue(), "Apiserver should be reachable passing from the SNI entrypoint in nodeport proxy")
-			ginkgo.By("tunneling requests using HTTP/2 CONNECT when no SNI is present e.g. pods relying on kubernetes service in default namespace")
-			// TODO: For sake of simplicity we are deploying an agent in the
-			// seed cluster. It would be better to create workers in the future for
-			// better coverage.
-			gomega.Expect(client.QueryApiserverVersion(agentConfig.GetKASHostPort(), true, options.kubernetesVersion, 5, 4)).To(gomega.BeTrue(), "Apiserver should be reachable passing from the SNI entrypoint in nodeport proxy")
-		})
-	})
-})
+			if err := clusterJig.CleanUp(ctx); err != nil {
+				t.Errorf("Failed to cleanup: %v", err)
+			}
+		}()
+	}
+}
